+++ packages/react-qr/src/Display.tsx
@@ -15,6 +15,7 @@ interface Props {
   size?: string | number;
   skipEncoding?: boolean;
   style?: React.CSSProperties;
+  timerDelay?: number;
   value: Uint8Array;
 }
 
@@ -30,7 +31,7 @@ interface TimerState {
   timerId: number | null;
 }
 
-const FRAME_DELAY = 2500;
+const DEFAULT_FRAME_DELAY = 2750;
 const TIMER_INC = 500;
 
 function getDataUrl (value: Uint8Array): string {
@@ -44,9 +45,9 @@ function getDataUrl (value: Uint8Array): string {
   return qr.createDataURL(16, 0);
 }
 
-function Display ({ className, size, skipEncoding, style, value }: Props): React.ReactElement<Props> | null {
+function Display ({ className, size, skipEncoding, style, timerDelay = DEFAULT_FRAME_DELAY, value }: Props): React.ReactElement<Props> | null {
   const [{ image }, setFrameState] = useState<FrameState>({ frameIdx: 0, frames: [], image: null, valueHash: null });
-  const timerRef = useRef<TimerState>({ timerDelay: FRAME_DELAY, timerId: null });
+  const timerRef = useRef<TimerState>({ timerDelay, timerId: null });
 
   const containerStyle = useMemo(
     () => createImgSize(size),
@@ -83,7 +84,7 @@ function Display ({ className, size, skipEncoding, style, value }: Props): React
       return newState;
     });
 
-    timerRef.current.timerId = window.setTimeout(nextFrame, FRAME_DELAY);
+    timerRef.current.timerId = window.setTimeout(nextFrame, timerRef.current.timerDelay);
 
     return (): void => {
       // eslint-disable-next-line react-hooks/exhaustive-deps